Move the legacy bundle specs into the project suite
They sat in pkg/service because that is where layer_test.go already opened a
whole service.Layer, and the project suite builds only a partial registry. An
external test package can import a package that depends on the one it tests, so
project_test opens the same layer without a cycle, and nothing forced the specs
a level up.

The fixtures move with them, and the specs join bundle_test.go, whose Import is
what they exercise. Its childrenOf reads the suite's shared services, so the
layer-scoped one is layerChildren.
