+++ core/pkg/service/project/testdata/legacy-projects/arc-editor-v8/Schematic.json
