+++ core/pkg/service/project/testdata/legacy-projects/hardware-v5/4ce50fe9-8682-4dcb-8e33-69b415538f92.json
