+++ core/pkg/service/project/testdata/legacy-projects/hardware-v5/4f2c7bca-cea7-483a-907f-f99ab608f3c7.json
