+++ core/pkg/service/project/testdata/legacy-projects/hardware-v5/4fa8328b-7d46-402d-a874-b7ab5eba6d4d.json
