+++ core/pkg/service/project/testdata/legacy-projects/hardware-v5/LAYOUT.json
