+++ core/pkg/service/project/testdata/legacy-projects/hardware-v5/e394d969-ae4c-413c-aa98-b8566841b252.json
