+++ core/pkg/service/project/testdata/legacy-projects/workspace-v4/6707c5bb-c4ec-45ce-9cc1-ed533511ccd5.json
