+++ core/pkg/service/project/testdata/legacy-projects/workspace-v4/LAYOUT.json
